Move governnace to tools (#27)
* chore: move governance to tools

* chore: replace super with crate
