+++ programs/nft-voter/src/tools/mod.rs
@@ -1,1 +1,2 @@
+pub mod governance;
 pub mod token_metadata;
