+++ programs/nft-voter/tests/cast_nft_vote.rs
@@ -1,5 +1,5 @@
 use crate::program_test::nft_voter_test::ConfigureCollectionArgs;
-use gpl_nft_voter::{error::NftVoterError, governance::VoterWeightAction};
+use gpl_nft_voter::{error::NftVoterError, tools::governance::VoterWeightAction};
 use program_test::{nft_voter_test::NftVoterTest, tools::assert_nft_voter_err};
 use solana_program_test::*;
 use solana_sdk::transport::TransportError;
